requirements.txt
pandas==0.20.3
idna==2.6
validators==0.12.0
domnibus==1.0.4
furl==1.0.1
Django==1.11.7
requests==2.18.4
dj_database_url==0.4.2
pyfav==0.1
tldextract==2.2.0
virustotal_api==1.1.7
Algorithmia==1.0.8
beautifulsoup4==4.6.0
certifi==2017.11.5
google_api_python_client==1.6.4
scikit_learn==0.19.1
dj-database-url==0.4.2
gunicorn==19.7.1
psycopg2==2.7.3.2
whitenoise==3.3.1
scipy